add template handling in main function
The main function's handling of templates has been refactored. Previously, if no template was found or it wasn't a string, an error message would be displayed and the function would return. Now, if no valid template is found, the default template from settings is used instead. The type assertion for 'template' has also been moved to where it's processed.